Update OS-Tan image URLs in README.md
Replaced direct image URLs with links to file pages on the Ostan Collections wiki. Updated the URL for the OS-Tan Windows XP image to a Pinterest link and the URL for the OS-Tan Windows Longhorn image to a different site. Also updated URLs for OS-Tan Windows 8 and OS-Tan Windows 10 images to their respective file pages on the Ostan Collections wiki.